Reusing existing PartitionedFileUtilsShims.
