+++ sql-plugin/src/main/spark400/scala/com/nvidia/spark/rapids/shims/PartitionedFileUtilsShim.scala
@@ -18,14 +18,14 @@
 /*** spark-rapids-shim-json-lines
 {"spark": "400"}
 spark-rapids-shim-json-lines ***/
-package org.apache.spark.sql.execution.rapids.shims
+package com.nvidia.spark.rapids.shims
 
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.execution.PartitionedFileUtil
 import org.apache.spark.sql.execution.datasources.{FileStatusWithMetadata, PartitionedFile}
 
-object PartitionedFileUtilShim extends PartitionedFileUtilShimBase {
+object PartitionedFileUtilsShim extends PartitionedFileUtilsShimBase {
 
   // In Spark 4.0, PartitionedFileUtil.splitFiles lost its `sparkSession` parameter.
   // This Spark-4.0+ shim ignores the `sparkSession` parameter.
